refactor: rename master page
